pyproject.toml
[project]
name = "pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "uv>=0.9.15",
    "pytorch",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "matplotlib>=3.10.7",
]
